+++ docs/changelog.md
@@ -7,6 +7,7 @@ Sprint 12 / May 10, 2020
 
 - added tests for Form, PieChart, PrivateRoute, and BarGraph components.
 - renamed Graph to BarGraph.
+- removed unnecessary initial time record.
 
 <hr />